Moved RunCTags out of legacy package
